+++ Sources/SuperwallKit/Superwall.swift
@@ -96,19 +96,7 @@ public final class Superwall: NSObject, ObservableObject {
   ///
   /// To learn more, see <doc:AdvancedConfiguration>.
   @Published
-  public var subscriptionStatus: SubscriptionStatus = .unknown {
-    didSet {
-      // `Self.shared` can't be accessed before `shared` has been configured.
-      if isConfigured == true,
-        !Self.shared.dependencyContainer.delegateAdapter.hasPurchaseController {
-        Logger.debug(
-          logLevel: .warn,
-          scope: .storeKitManager,
-          message: "Superwall.shared.subscriptionStatus set without active PurchaseController. Value may be later overwritten by Superwall. To manage subscriptionStatus on your own, or with subscription management platforms like RevenueCat and Qonversion, create a PurchaseController and pass it through to Superwall.configure(purchaseController:). Keep in mind, you will have to handle purchasing, restoring, and setting subscriptionStatus entirely on your own. If you are using RevenueCat, strongly consider using it in Observer Mode."
-        )
-      }
-    }
-  }
+  public var subscriptionStatus: SubscriptionStatus = .unknown
 
   /// A published property that is `true` when Superwall has finished configuring via
   /// ``configure(apiKey:purchaseController:options:completion:)-52tke``.